+++ .codebuild/buildspec.yml
@@ -14,6 +14,7 @@ env:
     CONFIRM_FIRST_ADDRESS: true
     VOIDED_TX_OFFSET: 20
     TX_HISTORY_MAX_COUNT: 50
+    CREATE_NFT_MAX_RETRIES: 3
     dev_DEFAULT_SERVER: "https://wallet-service.private-nodes.testnet.hathor.network/v1a/"
     dev_WS_DOMAIN: "ws.dev.wallet-service.testnet.hathor.network"
     dev_NETWORK: "testnet"
