+++ package.json
@@ -16,6 +16,7 @@
   "author": "André Abadesso <andre.abadesso@gmail.com>",
   "private": true,
   "devDependencies": {
+    "@types/jest": "^29.5.12",
     "@typescript-eslint/eslint-plugin": "^7.4.0",
     "@typescript-eslint/parser": "^7.4.0",
     "dotenv": "^16.4.5",
@@ -36,6 +37,7 @@
     "bip32": "^4.0.0",
     "bitcoinjs-lib": "^6.1.5",
     "bitcoinjs-message": "^2.2.0",
+    "jest": "^29.7.0",
     "tiny-secp256k1": "^2.2.3",
     "winston": "3.13.0"
   }
