+++ packages/common/__tests__/utils/alerting.utils.mock.ts
@@ -1,0 +1,4 @@
+export const mockedAddAlert = jest.fn();
+export default jest.mock('@src/utils/alerting.utils', () => ({
+  addAlert: mockedAddAlert.mockReturnValue(Promise.resolve()),
+}));
