+++ packages/common/package.json
@@ -1,12 +1,19 @@
 {
   "name": "@wallet-service/common",
   "packageManager": "yarn@4.1.0",
+  "scripts": {
+    "test": "jest --runInBand --collectCoverage --detectOpenHandles --forceExit"
+  },
   "peerDependencies": {
     "@aws-sdk/client-lambda": "3.540.0",
     "@hathor/wallet-lib": "0.39.0",
     "winston": "^3.13.0"
   },
   "devDependencies": {
-    "@types/node": "^20.11.30"
+    "@types/aws-lambda": "^8.10.136",
+    "@types/node": "^20.11.30",
+    "jest": "^29.6.4",
+    "ts-jest": "^29.1.2",
+    "typescript": "^5.4.3"
   }
 }
