+++ packages/daemon/__tests__/integration/balances.test.ts
@@ -42,6 +42,7 @@ import getConfig from '../../src/config';
 
 // @ts-ignore
 getConfig.mockReturnValue({
+  NETWORK: 'testnet',
   SERVICE_NAME: 'daemon-test',
   CONSOLE_LEVEL: 'debug',
   TX_CACHE_SIZE: 100,
@@ -85,6 +86,7 @@ describe('unvoided transaction scenario', () => {
   it('should do a full sync and the balances should match', async () => {
     // @ts-ignore
     getConfig.mockReturnValue({
+      NETWORK: 'testnet',
       SERVICE_NAME: 'daemon-test',
       CONSOLE_LEVEL: 'debug',
       TX_CACHE_SIZE: 100,
@@ -133,6 +135,7 @@ describe('reorg scenario', () => {
   it('should do a full sync and the balances should match', async () => {
     // @ts-ignore
     getConfig.mockReturnValue({
+      NETWORK: 'testnet',
       SERVICE_NAME: 'daemon-test',
       CONSOLE_LEVEL: 'debug',
       TX_CACHE_SIZE: 100,
@@ -181,6 +184,7 @@ describe('single chain blocks and transactions scenario', () => {
   it('should do a full sync and the balances should match', async () => {
     // @ts-ignore
     getConfig.mockReturnValue({
+      NETWORK: 'testnet',
       SERVICE_NAME: 'daemon-test',
       CONSOLE_LEVEL: 'debug',
       TX_CACHE_SIZE: 100,
