+++ packages/daemon/src/utils/wallet.ts
@@ -13,20 +13,22 @@ import {
   AddressBalance,
   AddressTotalBalance,
   DbTxOutput,
-  DecodedOutput,
   EventTxInput,
   EventTxOutput,
   StringMap,
-  TokenBalanceMap,
   TokenBalanceValue,
-  Transaction,
-  TxInput,
-  TxOutput,
-  TxOutputWithIndex,
   Wallet,
   WalletBalance,
   WalletBalanceValue,
 } from '../types';
+import {
+  DecodedOutput,
+  Transaction,
+  TxOutputWithIndex,
+  TxInput,
+  TxOutput,
+  TokenBalanceMap,
+} from '@wallet-service/common/src/types';
 import {
   fetchAddressBalance,
   fetchAddressTxHistorySum,
@@ -41,17 +43,6 @@ import logger from '../logger';
 import { stringMapIterator } from './helpers';
 
 /**
- * Checks if a given tokenData has any authority bit set
- *
- * tokenData merges two fields: first bit is the authority flag, while remaining
- * bits represent the token index. If the first bit is 0, this is a regular
- * output, if it's 1, it's an authority output
- */
-export const isAuthority = (tokenData: number): boolean => (
-  (tokenData & constants.TOKEN_AUTHORITY_MASK) > 0
-);
-
-/**
  * Prepares transaction outputs with additional metadata and indexing.
  *
  * This function expects a list  of EventTxOutput objects as inputs and an array
@@ -59,11 +50,11 @@ export const isAuthority = (tokenData: number): boolean => (
  * enhanced with additional data like the token it represents, its index in the
  * transaction, and its decoded information.
  *
- * @param outputs - An array of transaction outputs, each containing data like value, 
+ * @param outputs - An array of transaction outputs, each containing data like value,
  *                                    script, and token data.
- * @param tokens - An array of token identifiers corresponding to different tokens involved 
+ * @param tokens - An array of token identifiers corresponding to different tokens involved
  *                            in the transaction.
- * @returns - An array of outputs, each augmented with index and additional 
+ * @returns - An array of outputs, each augmented with index and additional
  *                                  metadata.
  */
 export const prepareOutputs = (outputs: EventTxOutput[], tokens: string[]): TxOutputWithIndex[] => {
@@ -270,13 +261,13 @@ export const unlockTimelockedUtxos = async (mysql: MysqlConnection, now: number)
 /**
  * Prepares transaction input data for processing or display.
  *
- * This function takes an array of EventTxInput objects and an array of token identifiers 
- * to prepare an array of TxInput objects. Each input is processed to include additional information 
+ * This function takes an array of EventTxInput objects and an array of token identifiers
+ * to prepare an array of TxInput objects. Each input is processed to include additional information
  * such as the token involved and the decoded output data.
  *
- * @param inputs - An array of transaction inputs, each containing data like 
+ * @param inputs - An array of transaction inputs, each containing data like
  *                                  transaction hash, index, and spent output information.
- * @param tokens - An array of token identifiers corresponding to different tokens involved 
+ * @param tokens - An array of token identifiers corresponding to different tokens involved
  *                            in the transaction.
  * @returns - An array of prepared inputs, each enriched with additional data.
  */
@@ -358,16 +349,16 @@ export const getTokenListFromInputsAndOutputs = (inputs: TxInput[], outputs: TxO
 
 /**
  * Validates the consistency of address balances.
- * 
- * This method is designed to validate that the sum of unlocked and locked balances 
- * for each address in a given set matches the corresponding total balance from the address's 
+ *
+ * This method is designed to validate that the sum of unlocked and locked balances
+ * for each address in a given set matches the corresponding total balance from the address's
  * transaction history.
  *
  * If any of these conditions are not met, the function will throw an assertion error, indicating a mismatch.
- * 
+ *
  * @param mysql - The MySQL connection object to perform database operations.
  * @param addresses - An array of addresses whose balances need to be validated.
- * @returns - The function returns a promise that resolves to void. It does not return 
+ * @returns - The function returns a promise that resolves to void. It does not return
  *                              any value but serves the purpose of validation.
  */
 export const validateAddressBalances = async (mysql: MysqlConnection, addresses: string[]): Promise<void> => {
