+++ packages/wallet-service/src/ws/connection.ts
@@ -20,7 +20,7 @@ import {
   initWsConnection,
   endWsConnection,
 } from '@src/redis';
-import { Severity } from '@src/types';
+import { Severity } from '@wallet-service/common/src/types';
 import { closeDbConnection, getDbConnection } from '@src/utils';
 import createDefaultLogger from '@src/logger';
 import { addAlert } from '@wallet-service/common/src/utils/alerting.utils';
